pyproject.toml
[tool.poetry]
name = "ehr-langchain-app"
version = "0.1.0"
description = ""
authors = ["Quyen Pham Khac <phamkhacquyen1996@gmail.com>"]
readme = "README.md"
packages = [{include = "ehr_langchain_app"}]

[tool.poetry.dependencies]
python = "^3.9"
langchain = "^0.1.3"
openai = "^1.9.0"
tiktoken = "^0.5.2"
chromadb = "^0.4.22"
python-dotenv = "^1.0.1"
langchain-openai = "^0.0.3"
pyboxen = "^1.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
